Fixed a little missconcept
By missconcept is meant that when a client sent a request towards the server from the contact us page, the server would send an email to that user. Now it is fixed, so when a client sends a request from contact us page it will be sent to the admin.